.gitmodules: For future use, add DrawThings gRPC protos as a submodule.
A gRPC proxy that wraps around Automatic1111 might be more efficient. It might be difficult to implement such a proxy, given the dependence on sha256s being used to address content-addressable files as inputs into ImageGenerationRequest, something not feasible with (near-)stock Automatic1111 image storage -- it would possibly require a new extension to even find the hashes for images we upload, depending on whether the temp images could reasonably be stored by the proxy.